requirements.txt
elasticsearch
spotipy
scikit-learn
pandas
pyspark
hashlib
implicit
tqdm
plotly
python-dotenv
argparse
django